feat: use decimal instead of float for money

+++ prisma/migrations/20260915000002_create_products/migration.sql
@@ -3,7 +3,7 @@ CREATE TABLE "products" (
     "id" SERIAL NOT NULL,
     "sku" TEXT NOT NULL,
     "name" TEXT NOT NULL,
-    "unit_price" DOUBLE PRECISION NOT NULL,
+    "unit_price" DECIMAL(12,2) NOT NULL,
     "weight_grams" INTEGER NOT NULL,
     CONSTRAINT "products_pkey" PRIMARY KEY ("id")
 );
